work on updating serialization
